Changed FDateTime UnixTimeStamp methods to use int64 rather than int32. Made relevant changes where the methods were used.
[CL 2500391 by Saul Abreu in Main branch]

+++ Engine/Source/Runtime/Core/Private/Internationalization/ICUText.cpp
@@ -27,7 +27,7 @@ FText FText::AsDate(const FDateTime& DateTime, const EDateTimeStyle::Type DateSt
 	checkf(I18N.IsInitialized() == true, TEXT("FInternationalization is not initialized. An FText formatting method was likely used in static object initialization - this is not supported."));
 	FCulturePtr Culture = TargetCulture.IsValid() ? TargetCulture : I18N.GetCurrentCulture();
 
-	int32 UNIXTimestamp = DateTime.ToUnixTimestamp();
+	int64 UNIXTimestamp = DateTime.ToUnixTimestamp();
 	UDate ICUDate = static_cast<double>(UNIXTimestamp) * U_MILLIS_PER_SECOND;
 
 	UErrorCode ICUStatus = U_ZERO_ERROR;
@@ -50,7 +50,7 @@ FText FText::AsTime(const FDateTime& DateTime, const EDateTimeStyle::Type TimeSt
 	checkf(I18N.IsInitialized() == true, TEXT("FInternationalization is not initialized. An FText formatting method was likely used in static object initialization - this is not supported."));
 	FCulturePtr Culture = TargetCulture.IsValid() ? TargetCulture : I18N.GetCurrentCulture();
 
-	int32 UNIXTimestamp = DateTime.ToUnixTimestamp();
+	int64 UNIXTimestamp = DateTime.ToUnixTimestamp();
 	UDate ICUDate = static_cast<double>(UNIXTimestamp) * U_MILLIS_PER_SECOND;
 
 	UErrorCode ICUStatus = U_ZERO_ERROR;
@@ -97,7 +97,7 @@ FText FText::AsDateTime(const FDateTime& DateTime, const EDateTimeStyle::Type Da
 	checkf(I18N.IsInitialized() == true, TEXT("FInternationalization is not initialized. An FText formatting method was likely used in static object initialization - this is not supported."));
 	FCulturePtr Culture = TargetCulture.IsValid() ? TargetCulture : I18N.GetCurrentCulture();
 
-	int32 UNIXTimestamp = DateTime.ToUnixTimestamp();
+	int64 UNIXTimestamp = DateTime.ToUnixTimestamp();
 	UDate ICUDate = static_cast<double>(UNIXTimestamp) * U_MILLIS_PER_SECOND;
 
 	UErrorCode ICUStatus = U_ZERO_ERROR;
